+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [0.0.13]
+* Documentation for `ScaffoldScope`
+
 ## [0.0.12]
 * Implement `Tooltip`
 * Add mouse cursors to help button, push button and `TextField`